requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
requests>=2.28.0
pandas>=2.0.0
numpy>=1.25.0
python-dotenv>=1.0.0
